IN=83997 - Add node factory interface and implementation to EJBQL compiler Add a node factory interface and an implementing class to the EJBQL compiler replacing classes ParseTreeConstructor and ParseTreeNodeConstructor.
The EJBQLParser class uses an instance of this interface to create an internal
representation of an EJBQL query. The parse tree is created bottom-up.
The EJBQL parse rules return the node representing the parsed EJBQL text.
This allows to pass any child nodes to the method creating the parent node.

Reviewed by: Tom Ware
Tests : EE QL, entity-persistence-tests, local tests