Issue Number: 6365261 Submitted By: Harpreet Sing Reviewed By : Satish Vishwanatham
Synopsis
Ebxml registry automatically saves a new version of classificationSchem when an Organization is saved. This results in multiple copies of classificationSchemes on the server. This creates problem when we query the registry to find a classificationScheme to unpublish the webservice.
A simple 1 line fix checks to see if it is a ebxml registry, then we do not publish a classification scheme and let saveOrganization to do the job.


Test Run: PE QL