Fixed https://glassfish.dev.java.net/issues/show_bug.cgi?id=118
Reviewed by : Sankar

Tests ran: QL, CTS-smoke, regression test for 6342889.

Description
-----------
As part of fix for 6342889, we introduced a new but dummy invocation for startups. This invocation was used as in indication for child threads to discard the parent's invocation stack. However, for persistence apps, when they load during startup, the information in the invocation stack was used. That was corrupted for that period of time, due to the dummy invocation.

The fix is to introduce a flag instead of a dummy invocation.