Fix for bug 6374631: NileBookStore, Unblocking keep-alive exception in grizzly connector
There is only one possibility of getting an NPE under load and when the file cache is enabled.  This it is produced by a thread race when the cache token expires and the static resource is being serviced simultaneously. Mainly, what happens is the static file is added to the cache for 30 seconds. When the delay occurs, the file is removed from the cache but there is a possibility that the cached ByteBuffer is still in use. This happens because when doing:

ByteBuffer1.slide()

which is extremely powerfull (since no sync required), a new ByteBuffer2 is created based on ByteBuffer1, called a ByteBuffer view. To avoid file locking problem on win32, the scheduler will set the ByteBuffer1 to null (the only way to remove the win32 lock) when the 30 seconds expires, but the SocketChannel may still using ByteBuffer2, thus the NPE.

- ALL passed
- NileBookStore works without exception.

Approved by Jerome Dochez (Bugswat)