Partial fix for 6374621: RichAccess, too many open files on imqbroker
Fix the Grizzly side. RichAccess application is used for stressing the server for 72 hours. Running this bug uncovered a file descriptor leak, which in turn is causing a memory leak. This commit redude the memory used by the file cache subsystems by:

- not initializing the header ByteBuffer unless a static resource is candidate to be added to the cache.
- not creating a FileCache object unless the resource is candidate to be cached
- do not cache FileCache token if monitoring is turned off
- Make sure socketChannel are deferenced when caching Algorithm and Handler so the VM can gc them.

No leak has been seens after making the above changes, but we are still seeing some funny file descriptor description:

7813: S_IFSOCK mode:0666 dev:323,0 ino:18992 uid:0 gid:0 size:0
      O_RDWR|O_NONBLOCK
    SOCK_STREAM
    SO_REUSEADDR,SO_SNDBUF(49152),SO_RCVBUF(49152)
    sockname: AF_INET 0.0.0.0  port: 0

so I'm not closing the bug as fixed until we find which component is creating this file decriptor leak.

- All QL passed
- All devtests/web passed
- All Servlet/Jsp TCKs passed
- All CTS smoke passed
- Static performance still good
- Maualy tested admin-gui

Reviewed by: Jan Luehe