Issue number:   212 Obtained from:  WSARC Submitted by:   Kedar Mhaswade Reviewed by:    Lloyd Chambers
This is the first of many checkins to be made to support asadmin login. The functionality being
checked in allows one to call the right routines to save the login information to ~/.asadminpass.

See https://glassfish.dev.java.net/javaee5/admin-infra/subprojects/asadminlogin/index.html

for details.

Note that nobody calls this API yet.