Fix for: 6360905 - App client behavior can change if current working directory changes
Problem: appclient script incorrectly includes the current working directory in the system classpath.  If the current directory subtree happens to contain a class or resource that should be provided by the the app client jar, for example, the client's behavior can change from one run to the next if the user changes the working directory.

This is a side-effect of allowing the user to specify APPCPATH as an environment variable.  The script adds the value of this env. var to the classpath but did not check to make sure the env. var. was defined before adding it to the classpath.

Fix: Make sure the env. var. is set before adding it to the classpath.

Reviewed by: Qingqing


Tests run: SQE quick look tests, deployment devtests, CTS smoke tests


Files affected:
  appserv-core/misc/templates/appclient.template
  appserv-core/misc/templates/appclient.bat.template