Partial fix for 6323557: admin GUI becomes non-responsive after adding a library
The AMX test suites is doing all king of strange call, like calling the WebCOntainer reconfig method but without creating first the WebContainer. So NPE are thrown because of that unexpected situation. I've added check (are they really needed?). Admin-gui still not working after, and now it fail with the security policy check. Re-assigning the bug to AMX team to make sure the WebContainer is started *before* any operation are made on the WebContainer.

- All QL passed.

-Reviewed by: Jan Luehe