Fix for 6374621: RichAccess, too many open files on imqbroker         6380645: Keep-Alive subsystems is sub-optimal
The keep-alive was using java.util.concurrent.ScheduledThreadPoolExecutor to periodically validate and times out idle connection. When ScheduledThreadPoolExecutor.schedule is invoked, it always creates an ScheduledThreadPoolExecutor$ScheduleEntry instance, and keep reference to this object forever whatever the scheduling of it has been cancelled or executed. Unless you explicitely invoke ScheduledThreadPoolExecutor.purge(), those objects will stay on the heap. For http keep-alive, the amount of such objects are very high, and cause a memory leak. Invoking the purge() method is not performant, so this commit remove all usage of ScheduledThreadPoolExecutor, and use the Selector.selectionKeys() queue to keep-alive connection. Everytime a connection is keep-alived, a Long is attached to the SelectionKey and compared everytime the Selector wakes up. No memory leaks have been observed, and raw throughput performance is much better with the new algorithm.

- All QL passed
- All devtest/web passed
- All Jsp/Servlet Tcks passed
- All CTS smoke passed
- RichAccess BigApp test is no longer leaking
- Ran trade2 benchmarks with performance improvement
- Manually tested admin-gui

Reviewed by: Scott Oaks