Fix for Issue 247: NPE when acessing web resource
This regression has started appearing the day we decided to add an Authenticator Valve for every application, security constraint defined or not (see 6202703). All Authenticator Valve are invoked before StandardContextValve, so the check for an invalid request was made after, hence the NPE. the bug is reproducabel everytime an invalid war file is deployed and exceptions happens during deployment. The HTTP mapper remove those invalid Wrapper, so we must look for null before invoking Authenticator Valve.

- All QL passed
- War attached to the bug are getting a 404 instead of 500.

Reviewed by: Jan Luehe