Fix for: 6382024 - jes5 stack reinstall till AM not working on AS
Problem: When AM installed with AS as part of Java ES, the uninstallation was incomplete.

It turns out that AM creates symbolic links in the AS domains/<domain-name> directory.  During uninstallation AM did not remove the link it had added.  The later delete-domain failed.  This was because the code that deletes files (and performs retries on Windows) checked for the existence of the file before deleting it.  It turns out that file.exists on a sym link checks for the pointed-to file or directory, whereas file.delete operates on the link itself.  Because the pointed-to directory had been removed already, the file.exists check failed so no delete was attempted by the FileUtils.deleteFile method.

The same problem affects 8.2 and 9.0.

Note that the Java ES team reports that the Windows installation does not use links.

Fix: Delete the requested file without checking for its existence first.  Revise the logging logic accordingly.

Reviewed by: Byron, Qingqing


Tests run: SQE PE quick look tests, deployment devtests, appclient devtests


File affected: appserv-commons/src/java/com/sun/enterprise/util/io/FileUtils.java