Issue number:   271 Submitted by: Qingqing Ouyang Reviewed by:  Hong Zhang
Modified Files:

M appserv-core/src/java/com/sun/enterprise/server/EJBModuleLoader.java

Description:

Bug fix for issue
https://glassfish.dev.java.net/issues/show_bug.cgi?id=271

For a standalone ejb module, both the BundleDescriptor (for the standalone ejb)
and the Application (skeleton wrapper for the ejb) have reference to the
classloader.

When we re-create the classloader for the application loading, only the
skeleton Application's ClassLoader gets reset, the BundleDescriptor does not.
Hence the problem.

Modified the EJBModuleLoader to also refresh the BundleDescriptor's reference
to the most current classloader.

Reviewer: Hong Zhang

Tests Ran:
1. PE QL/ deployment devtests / CTS smoke / ejb30/hello devtests
2. PE verified the app attached in the bug now runs.
3. Unit tested the modified ejb30 timer test