Bug#  - 6378363 (partial) post-PFD core EJB 3.0 spec updates Bug#  - 6298285 env-entry should not be injected when there is no env-entry-value subelement
Reviewers : QingQing Ouyang, Hong Zhang, Sheetal Vartak

Description :

1. Integrated latest ejb-jar_3_0.xsd

The only syntax change is the remove-if-exception element in
remove-method is optional instead of required.  The default
is no different than the default for @Remove when
retainIfException is not specified so there's no change in
behavior.

2. Bug 6298285

This is another case involving an env-entry that has no value specified in
the deployment descriptor.   Even if it has injection-targets specified,
injection shouldn't take place if there's no value.  Likewise, the
env-entry should not be visible in the component's java:comp/env.

3. The EJB 3.0 spec was clarified to state that timeout methods can
have any access type and can be defined on the bean superclass.

4. Added code to handle persistence-property overrides.

5. Removed check that  prevented ejb web service endpoint from
having TX attribute MANDATORY.   That restriction was just
lifted from the ejb spec.

Tests :

All ejb devtests passed.
All deployment devtests passed.
All sqe quicklook tests passed.

Related files:

  appserv-commons/schemas/ejb-jar_3_0.xsd
  appserv-commons/src/java/com/sun/enterprise/deployment/BundleDescriptor.java
  appserv-commons/src/java/com/sun/enterprise/deployment/EjbRemovalInfo.java
  appserv-commons/src/java/com/sun/enterprise/deployment/EnvironmentProperty.java
  appserv-commons/src/java/com/sun/enterprise/deployment/MethodDescriptor.java
  appserv-commons/src/java/com/sun/enterprise/deployment/annotation/handlers/AbstractEjbHandler.java
  appserv-commons/src/java/com/sun/enterprise/deployment/annotation/handlers/EntityManagerReferenceHandler.java
  appserv-commons/src/java/com/sun/enterprise/deployment/annotation/handlers/ResourceHandler.java
  appserv-commons/src/java/com/sun/enterprise/deployment/node/EnvEntryNode.java
  appserv-commons/src/java/com/sun/enterprise/deployment/util/EjbBundleValidator.java

  appserv-core/src/java/com/sun/ejb/containers/BaseContainer.java
  appserv-core/src/java/com/sun/ejb/containers/MessageBeanContainer.java
  appserv-core/src/java/com/sun/enterprise/naming/NamingManagerImpl.java