Issue number:   277 Obtained from:  asarch Submitted by:   Shing Wai Reviewed by:    Self
It seems that we can do something with -Djava.security.manager.

I have made changes such that the property -Dcom.sun.aas.installRoot=<the-appserver-install-root>
will now be passed to the app server VM from the launcher.

Other change is minor -- related to sorting the output.

Made sure that the server comes up properly and that the property does have the desired
value in server's VM invocation (java launch) command line.