Issue number:  6396744 Obtained from: - Submitted by: Sivakumar Thyagarajan Reviewed by:  Binod PG
As part of an earlier workaround for MQ bug 6336908 in AS, we made this changeset
http://fisheye5.cenqua.com/changelog/glassfish?cs=MAIN:sivakumart:20051219135654

This basically meant that we did not set ConnectionURL for EMBEDDED and LOCAL brokers in PE in the server side RA. However this meant that the connector descriptor bound in JNDI to be used in the appclient did not have the connectionURL property set either. Therefore, in the appclient, connectionURL is not set and hence MQ-RA while trying to create a connection configures MQ CF with a empty connection url [""], thereby attempting a connection to 7676.

Now that MQ bug # 6336908 is fixed, reverting this workaround in AS fixes the issue


Tests run: PE QL, Connector devtests, SQE JMS test