Fixed 6398929 ("Regression:Admin GUI is no longer function after deploy a jaxws 2.0 application") by backing out fix for 6377790 ("No user feedback is given when an exception is thrown during application deployment").
Proper fix for 6377790 would require extensive DOL changes, by
requiring the DOL to mark a webapp whose startup has failed as
disabled in domain.xml, so that this info would survive a server restart.

When the webcontainer disables a web module whose startup has thrown a
LifecycleException, that info currently does not survive a server
restart: It is stored in the web module itself, but during a restart,
the old web module instance is removed, and a new instance is
created. The new instance does not know that it is supposed to be
disabled.

If that information were persisted in the domain.xml, and made
available to the web container during a server restart, the web container
could skip starting any disabled modules.