Issue number:  400 Derby and db2 require a cast around parameter markers of certain operators if both the operands of the the operator are parameter markers. This integration adds functionality to add CAST
SQLCall is modified to delegate generating of parameter markers to DBPlatform
DB2Platform overrides this to generate appropriate CAST based on
ParameterExpression's  type.
The type is passed by ParameterNode to generated ParameterExpression

CONCAT also requires a VARCHAR cast. DB2Platform is modified for defn of CONCAT.
DerbyPlatform is modified to use the same definition of CONCAT as DB2

A test simpleConcatTestWithParameters() is added to JUnitEJBQLSimpleTestSuite to
exercise CONCAT with parameters.

Code reviewed by Peter and Gordon.

Tests run complete entity-persistence-tests on oracle, JUnitEJBQLSimpleTestSuite
on derby and db2