Description:
The reason the QL is failing is because someone broke the autodeploy.  I think it might be due to some changes in the admin code, though I don't understand it enough to tell which change caused this.

The fix, however, seem to be in the ApplicationsConfigMBean.java, which has not been modified for a few weeks.

It looks like there is a discrepancy in the name of the attribute for "http_listeners".  It should be "http_listeners", but the ACMB is looking for "http-listeners", which causes NPE, which bombs the autodeploy...

Reviewer: Hong Zhang

Tests Ran:
1. PE QL
2. manually tested autodeploy