Description:
Bug fix for
http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=6340561

Note that this doesnot address the request in the bug to validate the
"xmlns" namespace.  Though that part is in debat if we should even do
it.

What this fix does address is the fact that we now is not following
the configuration for the validationLevel for application.xml.  This
is because the application.xml is read separately in the
J2EEModuleExploder.  The fix is to make sure the validationLevel is
set according to the configuration in domain.xml (if available).

Reviewer: Hong Zhang

Tests Ran:
1. PE QL
2. manaully tested verifier standalone case, with rogue applications
3. deployment devtests
4. devtsts/ejb/ejb30/hello