Issue number: 6396744 Obtained from: - Submitted by:  Sivakumar Thyagarajan Reviewed by:   Binod PG
The earlier fix for this bug resulted in 5 CTS smoke test failures and was backed out.
On analysis of the 5 CTS test failures it was understood that the AddressList property should be set only for REMOTE and/or when additional JMS hosts are available. The default LOCAL and EMBEDDED modes in the server should not set AddressList. Setting an AddressList in the LOCAL and EMBEDDED modes results in the MQ-RA adding an additional host to the address list and the broker throws an exception says multiple hosts are not allowed for PE license. Fix is to set AddressList property only for REMOTE and/or when additional JMS hosts are available.

Tests run
Linux : PE QL, CTS Smoke, connector devtests and SQE JMS  on PE on glassfish image
Ran PE QL and CTS smoke on installer image
Manually ensured that AddressList was set only for REMOTE mode