Issue number:   319 Obtained from:  Issue Tracker Submitted by:   Kedar Reviewed by:    Kedar
Provide the backend support for connectors to get the password from the
domain-passwords store through alias.
Without this, I don't know how it would have *ever* worked.

Will transfer the bug to Binod to add an appropriate call to the API.

No QL impact as the code is *not* yet called. Thorough manual testing.