Changed
  setInitParameters(Map initParms)

to

  setInitParameters(Map<String, Object> initParms)

to make it consistent with the signature of the setInitParameters method
of the javax.servlet.jsp.tagext.TagLibraryValidator superclass, which
already uses generics.

The JSTL related part of this change had already been communicated to the
JSTL EG and is reflected in the JSTL 1.2 MR changelog, but apparently the
API was never updated.