Description:
Bug fixes for the following.
http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=6254405
http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=6399044

Upon application loading failures, we explicitly register the error
with the AdminEventResults instead of throwing an AdminEventException.
This way the server restartRequired flag won't be set to true.

Also double check the descriptor object for null.  If the
application loading fails, the descriptor is cleared out.  We
need to guard this case during redeployment.

Reviewer: Hong Zhang (deployment)
          Alex K. (admin)

Tests Ran:

1. PE QL
2. deployment devtests
3. ejb30/hello devtests
4. manually tested the following usecase:

    deploy/redeploy/redeploy of an ejb jar that fails to load
    disable/enable of an ejb jar that fails to load
    deploy/redeploy/redeploy of an ear file that fails to load
    disable/enable of an ear file that fails to load
    checked restart flag after each operation.