Fix for bug 6403933: java.nio.Selector doesn't block on Selector.select(timeout)
Instead of clearing the SelectionKey set, cancel the temporary key and wakeup the Selector. The previous implementation is supposed to work, but unfortunaly in some unpredicted case, the Selector isn't blocking on select, causing an 'broken pipe' exception on the client side. Adding this workaround fix the problem.

+ All QL passed
+ All devtests/web passed (the 'broken pipe' problem was only reproducable using that test)

Workaround proposed by Charlie Hunt.
Reviewed by Scott Oaks