Issue number:  6402313 Obtained from: - Submitted by:  Sivakumar Thyagarajan Reviewed by:   Binod PG, Kedar Mhaswade/Sreenivas Munnangi Fix for issue 6402313
This was a regression in add-resources from 8.2. When a resource creation failed in the resource list specified via a resources.xml file, we continue creation of other resources specified in 8.2 Made changes to ResourcesMBean to achieve this. The change also ensures that the API for teh deployment team [createResourceAndResourceRef for sun-resources.xml processing] does not break with these changes.

Tests run
Linux : PE QL, Manually ensured that testcase provided with the bug report passes