Bug#  - Issue 525 : Error report when DB is not running
Reviewers : QingQing Ouyang

Description :

Fixed delegateInfo initialization bug in EntityManagerWrapper.
Moved the initialization of delegateInfo before the entityManagerFactory
logic so that it will be initialized even if an IllegalStateException
is thrown.

Tests :

Verified that IllegalStateException is thrown instead of
NullPointerException if physical EntityManagerFactory resolution
fails as a result of EntityManagerWrapper.createQuery().

All ejb devtests passed.
All Quicklook tests passed.

Related files:

  appserv-core/src/java/com/sun/enterprise/util/EntityManagerWrapper.java