Partial fix for 6415801 - SEVERE: Could not expand entry com\\\\sun\\\\ts\\\\tests\\\\jws\\\\webservice\\\\webservice3\\\\client\\ ... Fix for 6412114 - Deployment fails with the error "Windows file path is too long"
(Porting the fix from 9.0 FCS to 9.0 main trunk)

Problem: Windows file path length check prevents expansion of archives and is overly restrictive

Logic was inserted in an earlier app server release to check that the path length of an output file during jar expansions did not exceed the Windows limit.  An underlying problem in Java SE has been fixed in 1.5.0_06.  This check-in removes the logic that checks the path length on Windows.  On Java SE releases prior to _06 users whose paths exceed the Windows limit should expect to see an IOException that the path syntax is incorrect.

Reviewed by: Hong


Tests:
  SQE quick-look tests, deployment dev tests, CTS smoke tests, CTS TCK tests that failed (jws/webservice/webservice3 and webservices12/servlet)


appserv-commons/src/java/com/sun/enterprise/util/io
  FileUtils.java