Fix for Issue 601 - deployment on localized windows has problem
Problem: Deployment fails when the Windows temp directory contains a multibyte character

Diagnosis: The upload servlet responds to the client with the server-side file path to the uploaded file.  The client sends this back to the server in a subsequent request.  The servlet's response, though, did not specify UTF-8 encoding.  The string received and resent by the client did not, then, correspond exactly to the correct path on the server, and the server could not find the uploaded file to be deployed.

Fix: Specify UTF-8 as the encoding on the response to the client from the upload servlet.  Also, the client-side FileUploadUtil class specified UTF8 instead of UTF-8 in two places.  This seemed not to have a direct effect on this problem but was cleaned up anyway.

Reviewed by: Hong


Tests:

PE SQE quick-looks; deployment devtests


Code changes:

appserv-core/src/java/com/sun/enterprise/admin/server/core/servlet/
 UploadServlet.java

appserv-commons/src/java/com/sun/enterprise/deployment/util/
 FileUploadUtil.java