Issue number:  635 Obtained from: Issue tracker Reviewed by:   Gordon Yorke
IN=95500 - Prevent NullPointerException for Object typed write lock (version) fields

For instances having an Object typed version field, e.g. Integer, a
NullPointerException might be thrown when comparing versions in
ObjectChangeSet#compareWriteLockValues.
Added a new method setting the initialWriteLockValue in ObjectChangeSet, which
is called when an object is first inserted to the database in
VersionLockingPolicy#setupWriteFieldsForInsert. This makes sure, that the value
initialWriteLockValue is always set.

Tested by Quicklook and entity-persistence-tests.