Entity-Persistence update
Submitted by Tom Ware
Reviewed by TopLink Team

Tested with QuickLook, entity-persistence-tests and Oracle-internal tests

- ailitche_ri_5237881_060620  Update to fix for Oracle bug 5237881:
CLASS FILE IN NON-PACKAGE DIRECTORY CAUSES EJB 3.0 W/ JPA TO FAIL - This
is related to our PersistenceUnitProcessor and how it reacts to class
files it finds that are not actually on the classpath
- tware_10essentials_gf711_valueholder_instantiation_exception_060615
Glassfish Bug 711 - better exception when crossing an uninstantiated
valueholder post serialization
- tware_10_5031472_same_name_expression_060613  Fix for Oracle bug
5031472 - This bug occured when writing a TopLink expression that
traversed and OneToOne mapping that had the same name as a mapping in
the destination object and also traversed the mapping with the same name
in the destination object.  To see the bug, the foreign key for the
destination object would have to be part of the Primary Key for that object.
gyorke_fix-joined-attributes-not-cloned-correctly_060612 and
gyorke_fix-null-pointer-in-ob-build-attr-into-obj_060613
joined-attributes-not-cloned-correctly

: ----------------------------------------------------------------------