Issue number: entity-persistence-update Obtained from: TopLink Submitted by: Tom Ware Reviewed by:  TopLink Team
rity of this update is a complete refactor of the XML and
Annotations processing

framework.  It is broken up into several transactions.  Details are
listed below with a transaction identifier that can be matched with the
attached diff file.

ORM Refactoring
- gpelleti_ri_metadataprocessingfixes_060627  Fixes a class cast exception.
- gpelleti_ri_ormfixes_060627  Fixes some sequencing metadata processing
errors.
- gpelleti_ri_metadataprocessingfixes_060626  Fix for embeddable
superclasses that are defined in a seperate XML file from where the
entities that subclass them are defined.
- gpelleti_ri_xmlannotationinhertiancemixfix_060624  OR metadata
processing re-work. Fix for mixing inheritance entity definitions in
annotations and xml.
- gpelleti_ri_ormfixes_060621  OR metadata processing re-work.

Oracle bug  5170395
- ailitche_ri_5170395_060621  Fixed Oracle bug 5170395 - Sequencing
exception when running for the first time on a new schema