Issue number:  810 Obtained from: Issue tracer Submitted by: Reviewed by:   Guy Pelletier IN=97474 - Prevent NPE after MetadataProject#getDescriptor is called
After initialization, method MetadataProject#getDescriptor(Class cls) is
supposed to never return null. In case of misconfiguration, i.e. a class is not
listed in persistence.xml and exclude-unlisted-classes equals true,
getDescriptor actually returns null, causing a NullPointerException later on.

Now getDescriptor throws a ValidationException if the descriptor for any
argument not equals null is not found.