Bug#  835 - Container managed Entity Manager does not honor maxResults Parameter in Queries
Synopsis : Queries created through non-transactional access of a
container-managed entity manager not honoring maxResults/firstResults
api for the first execution.

Reviewers : Mahesh Kannan

Description :

Backporting to 9.0 UR1.

QueryWrapper, which is only used when a query is created through
non-transactional use of a container-managed entity manager,
was not delegating the setMaxResults/setFirstResult call
to the existing QueryDelegate.  Therefore the first call to
getResultList was not including that setting.  Added the delegate
calls to the QueryWrapper.

Added assertions to devtests/ejb/ejb30/persistence/context

All ejb devtests passed.
All sqe quicklook tests passed.

Related files:
  appserv-core/src/java/com/sun/enterprise/util/QueryWrapper.java