Issue number:  6443036 Obtained from: - Submitted by:  Kshitiz Saxena Reviewed by:  Sivakumar Thyagarajan
In ResourceInstaller, the standard activation-config  was taken as activation-config for MDB. It was not taking runtime activation-config into consideration. We needed to merge the two to get final activation-config. This was being done in
ConnectorMessageBeanClient. So extracted the method to RARUtils class and used it in both of them.

Tests run: QL, connector devtests, CTS Smoketest