Issue number:  6430394 and 6436024 Obtained from: - Submitted by:  Sivakumar Thyagarajan Reviewed by:   Binod PG
6430394

Changed files:
appserv-core/src/java//com/sun/enterprise/resource/ResourceManagerImpl.java

Description of fix: at certain points, when a XA resource could not be
registered as part of a component's resource list, due to a failure
while enlisting the XAResource in the transaction, we weren't cleaning
up the resource from the component list. This resulted in the same
XAResource being reused when the component(MDB) is reused. The current
fix ensures that the resource is cleaned from the component list when
there is failure elisting the XA resource in the transaction

6436024
Since "useExternalRMIRegistry" is not required for REMOTE configured RAs
the check at
appserv-core/src/java/com/sun/enterprise/connectors/system/ActiveJmsResourceAdapter.shouldUseExternalRmiRegistry()
need not be done for REMOTE.

Tests run : PE QL, Connector dev tests, manually tested that there is no
delay when we have two domains