Added javadocs and NEWS tag
[NEWS:webcontainer 587] Fixed WebappClassLoader memory leak
Fix removes all bean classes loaded by a webapp's WebappClassLoader from
the caches maintained by javax.el.BeanELResolver when the webapp is being
undeployed (that is, during WebappClassLoader.stop()).
As of now, the cache needs to be accessed reflectively, since
javax.el.BeanELResolver does not expose any APIs that would allow it
to be notified of a webapp being undeployed. Hopefully, an appropriate
API can be added in the next version of the JSP spec.