Fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=904 ("extra-class-path attribute in <class-loader> should allow path relative to docroot")
Tests run:
- QL/PE

[NEWS:webcontainer 904] Relative extra class paths are now interpreted as relative to a context's docroot
Any relative paths listed as the value of the extra-class-path
attribute of the class-loader element in sun-web.xml are now interpreted
as relative to a context's docroot.
This is certainly more useful than the existing behaviour, which has
interpreted them as relative to $INSTANCE_ROOT/config.
This is not going to break any backwards compatibility, because the
current documentation of this attribute only mentions that it
"specifies additional classpath settings for this web module", without
explaining how relative paths are to be treated.